+++ test/utils/markup.test.js
@@ -1,7 +1,7 @@
 import { html } from 'lit-html';
 import { BaElement } from '../../src/modules/BaElement';
 import { MvuElement } from '../../src/modules/MvuElement';
-import { calculateWorkingArea, decodeHtmlEntities, findAllByAttribute, forEachByAttribute, TEST_ID_ATTRIBUTE_NAME } from '../../src/utils/markup';
+import { decodeHtmlEntities, findAllByAttribute, forEachByAttribute, TEST_ID_ATTRIBUTE_NAME } from '../../src/utils/markup';
 import { TestUtils } from '../test-utils';
 
 class MvuElementParent extends MvuElement {
@@ -180,125 +180,6 @@ describe('markup utils', () => {
 		});
 	});
 
-	describe('calculateWorkingArea', () => {
-		const getBase = () => {
-			return { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 1000, height: 1000 }) };
-		};
-
-		it('calculates an workingArea besides a leftSidePanel', () => {
-			const leftSidePanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 100, height: 1000 }) };
-			const dirtyElements = [leftSidePanelMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(100);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(1000);
-		});
-
-		it('calculates an workingArea besides a leftSidePanel and a bottomPanel', () => {
-			const leftSidePanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 100, height: 1000 }) };
-			const bottomPanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 100, y: 900, width: 800, height: 100 }) };
-			const dirtyElements = [leftSidePanelMock,	bottomPanelMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(100);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(900);
-			expect(result.bottom).toBe(900);
-		});
-
-		it('calculates an workingArea besides a leftSidePanel and a rightSideBottomPanel', () => {
-			const leftSidePanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 100, height: 1000 }) };
-			const rightSideBottomPanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 800, y: 900, width: 200, height: 100 }) };
-			const dirtyElements = [leftSidePanelMock,	rightSideBottomPanelMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(100);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(800);
-			expect(result.bottom).toBe(1000);
-		});
-
-
-		it('calculates an workingArea besides a topPanel and a bottomPanel', () => {
-			const topPanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 1000, height: 300 }) };
-			const bottomPanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 700, width: 1000, height: 300 }) };
-			const dirtyElements = [topPanelMock,	bottomPanelMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(0);
-			expect(result.top).toBe(300);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(700);
-		});
-
-		it('calculates an workingArea besides a leftSidePanel and a overlappingLeftSidePanel', () => {
-			const leftSidePanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 100, height: 1000 }) };
-			const overlappingLeftSidePanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 200, height: 500 }) };
-			const dirtyElements = [leftSidePanelMock,	overlappingLeftSidePanelMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(200);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(1000);
-		});
-
-		it('calculates an workingArea besides a centered element', () => {
-			const centeredElementMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 450, y: 450, width: 100, height: 100 }) };
-			const dirtyElements = [centeredElementMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(550);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(1000);
-		});
-
-
-		it('calculates an workingArea besides an empty element', () => {
-			const emptyElementMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: 0, y: 0, width: 0, height: 0 }) };
-			const dirtyElements = [emptyElementMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(0);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(1000);
-		});
-
-		it('calculates an workingArea besides an disjoint element', () => {
-			const emptyElementMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: -300, y: 0, width: 300, height: 1000 }) };
-			const dirtyElements = [emptyElementMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(0);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(1000);
-		});
-
-		it('calculates an workingArea besides an partiallyOverlappingLeftSidePanel', () => {
-			const partiallyOverlappingLeftSidePanelMock = { getBoundingClientRect: () => DOMRect.fromRect({ x: -50, y: 0, width: 300, height: 1000 }) };
-			const dirtyElements = [partiallyOverlappingLeftSidePanelMock];
-
-			const result = calculateWorkingArea(getBase(), dirtyElements);
-
-			expect(result.left).toBe(250);
-			expect(result.top).toBe(0);
-			expect(result.right).toBe(1000);
-			expect(result.bottom).toBe(1000);
-		});
-	});
 
 	describe('forEachByAttribute', () => {
